refactor reward views
