+++ bootstrap/src/test/java/onlydust/com/marketplace/api/bootstrap/it/api/InvoicesApiIT.java
@@ -999,7 +999,7 @@ public class InvoicesApiIT extends AbstractMarketplaceApiIT {
         assertThat(invoiceOnCompany.status()).isEqualTo(InvoiceEntity.Status.DRAFT);
         assertThat(invoiceOnCompany.data().rewards().stream().anyMatch(r -> r.id().equals(UUID.fromString(rewardId)))).isTrue();
         var reward = entityManagerFactory.createEntityManager().find(RewardEntity.class, UUID.fromString(rewardId));
-        assertThat(reward.invoice().id().toString()).isEqualTo(invoiceOnCompanyId.getValue());
+        assertThat(reward.invoiceId().toString()).isEqualTo(invoiceOnCompanyId.getValue());
 
         // Then switch billing profile type
         final var ownerId = UserId.of(antho.user().getId());
@@ -1042,7 +1042,7 @@ public class InvoicesApiIT extends AbstractMarketplaceApiIT {
 
         // Check that the reward is part of the approved invoice
         reward = em.find(RewardEntity.class, UUID.fromString(rewardId));
-        assertThat(reward.invoice().id().toString()).isEqualTo(invoiceId.getValue());
+        assertThat(reward.invoiceId().toString()).isEqualTo(invoiceId.getValue());
 
         // Switch back to company billing profile type
         payoutPreferenceFacadePort.setPayoutPreference(PROJECT_ID, BillingProfile.Id.of(companyBillingProfileId), ownerId);
@@ -1061,7 +1061,7 @@ public class InvoicesApiIT extends AbstractMarketplaceApiIT {
 
         // Now check that the reward is still part of the approved invoice
         reward = em.find(RewardEntity.class, UUID.fromString(rewardId));
-        assertThat(reward.invoice().id().toString()).isEqualTo(invoiceId.getValue());
+        assertThat(reward.invoiceId().toString()).isEqualTo(invoiceId.getValue());
         em.close();
     }
