+++ bootstrap/src/test/java/onlydust/com/marketplace/api/bootstrap/it/api/ProjectsGetContributorsApiIT.java
@@ -1767,25 +1767,25 @@ public class ProjectsGetContributorsApiIT extends AbstractMarketplaceApiIT {
         final UUID projectId = UUID.fromString("f39b827f-df73-498c-8853-99bc3f562723");
 
         final var reward1 = rewardRepository.findById(UUID.fromString("8fe07ae1-cf3b-4401-8958-a9e0b0aec7b0")).orElseThrow();
-        reward1.currency(currencyRepository.findByCode("ETH").orElseThrow());
+        reward1.currencyId(currencyRepository.findByCode("ETH").orElseThrow().id());
         reward1.amount(BigDecimal.valueOf(20.5));
         rewardRepository.save(reward1);
         rewardStatusRepository.save(rewardStatusRepository.findById(reward1.id()).orElseThrow().amountUsdEquivalent(BigDecimal.valueOf(31426.5)));
 
         final var reward2 = rewardRepository.findById(UUID.fromString("e1498a17-5090-4071-a88a-6f0b0c337c3a")).orElseThrow();
-        reward2.currency(currencyRepository.findByCode("APT").orElseThrow());
+        reward2.currencyId(currencyRepository.findByCode("APT").orElseThrow().id());
         reward2.amount(BigDecimal.valueOf(2000));
         rewardRepository.save(reward2);
         rewardStatusRepository.save(rewardStatusRepository.findById(reward2.id()).orElseThrow().amountUsdEquivalent(BigDecimal.valueOf(1120)));
 
         final var reward3 = rewardRepository.findById(UUID.fromString("40fda3c6-2a3f-4cdd-ba12-0499dd232d53")).orElseThrow();
-        reward3.currency(currencyRepository.findByCode("OP").orElseThrow());
+        reward3.currencyId(currencyRepository.findByCode("OP").orElseThrow().id());
         reward3.amount(BigDecimal.valueOf(450));
         rewardRepository.save(reward3);
         rewardStatusRepository.save(rewardStatusRepository.findById(reward3.id()).orElseThrow().amountUsdEquivalent(BigDecimal.valueOf(643.5)));
 
         final var reward4 = rewardRepository.findById(UUID.fromString("5b96ca1e-4ad2-41c1-8819-520b885d9223")).orElseThrow();
-        reward4.currency(currencyRepository.findByCode("STRK").orElseThrow());
+        reward4.currencyId(currencyRepository.findByCode("STRK").orElseThrow().id());
         reward4.amount(BigDecimal.valueOf(500000));
         rewardRepository.save(reward4);
         rewardStatusRepository.save(rewardStatusRepository.findById(reward4.id()).orElseThrow().amountUsdEquivalent(null));
