+++ infrastructure/postgres-adapter/src/main/java/onlydust/com/marketplace/api/postgres/adapter/entity/write/BatchPaymentRewardEntity.java
@@ -2,8 +2,9 @@ package onlydust.com.marketplace.api.postgres.adapter.entity.write;
 
 import lombok.*;
 import lombok.experimental.Accessors;
-import onlydust.com.marketplace.accounting.domain.model.Payment;
 import onlydust.com.marketplace.accounting.domain.model.PayableReward;
+import onlydust.com.marketplace.accounting.domain.model.Payment;
+import onlydust.com.marketplace.api.postgres.adapter.entity.read.RewardViewEntity;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.*;
@@ -32,9 +33,10 @@ public class BatchPaymentRewardEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "batchPaymentId", referencedColumnName = "id", insertable = false, updatable = false)
     BatchPaymentEntity batchPayment;
+
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "rewardId", referencedColumnName = "id", insertable = false, updatable = false)
-    RewardEntity reward;
+    RewardViewEntity reward;
 
     @EqualsAndHashCode
     @AllArgsConstructor
